Rename `@inngest/workflow-kit`; fix some example imports
