+++ examples/nextjs-blog-cms/package.json
@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@inngest/workflow": "workspace:*",
+    "@inngest/workflow-kit": "workspace:*",
     "@mdx-js/mdx": "^3.0.1",
     "@radix-ui/react-checkbox": "^1.1.1",
     "@radix-ui/react-icons": "^1.3.0",
